+++ sdk/storage/storage-queue/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 12.2.1 (Unreleased)
 
+- Updated Azure Storage Service API version to 2020-04-08.
 
 ## 12.2.0 (2020-11-10)